move userFindByEmail functions into own file to avoid dependency
