+++ packages/lesswrong/server/authenticationMiddlewares.ts
@@ -17,7 +17,7 @@ import { forumTypeSetting } from '../lib/instanceSettings';
 import { userFromAuth0Profile } from './authentication/auth0Accounts';
 import { captureException } from '@sentry/core';
 import moment from 'moment';
-import { userFindOneByEmail, usersFindAllByEmail } from "../lib/collections/users/helpers";
+import {userFindOneByEmail, usersFindAllByEmail} from "../lib/collections/users/commonQueries";
 
 /**
  * Passport declares an empty interface User in the Express namespace. We modify
